Rename CompletionStatus to PaymentResult
