+++ stripe/src/main/java/com/stripe/android/paymentsheet/PaymentSheet.kt
@@ -2,10 +2,8 @@ package com.stripe.android.paymentsheet
 
 import android.content.Intent
 import android.os.Bundle
-import android.os.Parcelable
 import androidx.activity.ComponentActivity
 import androidx.core.os.bundleOf
-import com.stripe.android.model.PaymentIntent
 import com.stripe.android.view.ActivityStarter
 import kotlinx.android.parcel.Parcelize
 
@@ -36,34 +34,15 @@ internal class PaymentSheet internal constructor(
         PaymentSheetActivityStarter.Args.Guest(clientSecret)
     )
 
-    fun confirm(activity: ComponentActivity, callback: (CompletionStatus) -> Unit) {
+    fun confirm(activity: ComponentActivity, callback: (PaymentResult) -> Unit) {
         // TODO: Use ActivityResultContract and call callback instead of using onActivityResult
         // when androidx.activity:1.2.0 hits GA
         PaymentSheetActivityStarter(activity)
             .startForResult(args)
     }
 
-    internal sealed class CompletionStatus : Parcelable {
-        @Parcelize
-        data class Succeeded(
-            val paymentIntent: PaymentIntent
-        ) : CompletionStatus()
-
-        @Parcelize
-        data class Failed(
-            val error: Throwable,
-            val paymentIntent: PaymentIntent?
-        ) : CompletionStatus()
-
-        @Parcelize
-        data class Cancelled(
-            val mostRecentError: Throwable?,
-            val paymentIntent: PaymentIntent?
-        ) : CompletionStatus()
-    }
-
     @Parcelize
-    internal data class Result(val status: CompletionStatus) : ActivityStarter.Result {
+    internal data class Result(val status: PaymentResult) : ActivityStarter.Result {
         override fun toBundle(): Bundle {
             return bundleOf(ActivityStarter.Result.EXTRA to this)
         }
